+++ docs/examples/aliases/delete.ts
@@ -1,0 +1,16 @@
+import { Client } from '@robgietema/nick';
+
+const cli = Client.initialize({ apiPath: 'http://localhost:8080' });
+const login = await cli.login({ username: 'admin', password: 'admin' });
+
+const { data } = await cli.deleteAliases({
+  token: login.data.token,
+  path: '/news',
+  data: {
+    items: [
+      {
+        path: '/news-items',
+      },
+    ],
+  },
+});